# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: MFK — Smart Vehicle Diagnostics Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{vehicleId}:
    parameters:
      - in: path
        name: vehicleId
        required: true
        schema: { type: string }
    get:
      operationId: getVehicle
      tags: [vehicles]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleDetail"
    patch:
      operationId: updateVehicle
      tags: [vehicles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVehicleBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      responses:
        "204":
          description: Deleted
...
  /dtc:
    get:
      operationId: listDtcCodes
      tags: [dtc]
      parameters:
        - in: query
          name: vehicleId
          schema: { type: string }
        - in: query
          name: status
          schema:
            type: string
            enum: [active, cleared, pending]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DtcCode"
  /dtc/{dtcId}:
    parameters:
      - in: path
        name: dtcId
        required: true
        schema: { type: string }
    get:
      operationId: getDtcCode
      tags: [dtc]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DtcCode"
  /dtc/{dtcId}/clear:
    parameters:
      - in: path
        name: dtcId
        required: true
        schema: { type: string }
    post:
      operationId: clearDtcCode
      tags: [dtc]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DtcCode"
...
  /workshops:
    get:
      operationId: listWorkshops
      tags: [workshops]
      parameters:
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: service
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workshop"
  /workshops/{workshopId}:
    parameters:
      - in: path
        name: workshopId
        required: true
        schema: { type: string }
    get:
      operationId: getWorkshop
      tags: [workshops]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkshopDetail"

  # ========== BOOKINGS ==========
...
  /ai/interpret-dtc:
    post:
      operationId: interpretDtc
      tags: [ai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterpretDtcBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DtcInterpretation"

  # ========== SUBSCRIPTIONS ==========
...
components:
  schemas:
...
    Vehicle:
      type: object
      required: [id, make, model, year, healthScore]
      properties:
        id: { type: string }
        vin: { type: string, nullable: true }
        make: { type: string }
        model: { type: string }
        year: { type: integer }
        plateNumber: { type: string, nullable: true }
        nickname: { type: string, nullable: true }
        odometerKm: { type: integer }
        fuelType:
          type: string
          enum: [petrol, diesel, hybrid, ev]
        engineCc: { type: integer, nullable: true }
        adapterMac: { type: string, nullable: true }
        healthScore: { type: integer }
        imageUrl: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    VehicleDetail:
      allOf:
        - $ref: "#/components/schemas/Vehicle"
        - type: object
          properties:
            activeDtcCount: { type: integer }
            upcomingMaintenanceCount: { type: integer }
            lastSessionAt: { type: string, format: date-time, nullable: true }
            totalSessions: { type: integer }
            isPaired: { type: boolean }
    CreateVehicleBody:
      type: object
      required: [make, model, year, fuelType]
      properties:
        vin: { type: string }
        make: { type: string }
        model: { type: string }
        year: { type: integer }
        plateNumber: { type: string }
        nickname: { type: string }
        odometerKm: { type: integer }
        fuelType:
          type: string
          enum: [petrol, diesel, hybrid, ev]
        engineCc: { type: integer }
    UpdateVehicleBody:
      type: object
      properties:
        nickname: { type: string }
        plateNumber: { type: string }
        odometerKm: { type: integer }
...
    DtcCode:
      type: object
      required: [id, code, status, severity, vehicleId]
      properties:
        id: { type: string }
        sessionId: { type: string, nullable: true }
        vehicleId: { type: string }
        vehicleMake: { type: string, nullable: true }
        vehicleModel: { type: string, nullable: true }
        code: { type: string }
        status:
          type: string
          enum: [active, cleared, pending]
        severity:
          type: string
          enum: [low, medium, high, critical]
        descriptionEn: { type: string, nullable: true }
        descriptionAr: { type: string, nullable: true }
        possibleCauses:
          type: array
          items: { type: string }
        estimatedCostMin: { type: integer, nullable: true }
        estimatedCostMax: { type: integer, nullable: true }
        recommendedAction:
          type: string
          nullable: true
          enum: [drive_now, schedule_week, monitor]
        actionReasonAr: { type: string, nullable: true }
        detectedAt: { type: string, format: date-time }
        clearedAt: { type: string, format: date-time, nullable: true }
...
    Workshop:
      type: object
      required: [id, name, nameAr, city, rating]
      properties:
        id: { type: string }
        name: { type: string }
        nameAr: { type: string }
        phone: { type: string, nullable: true }
        city: { type: string }
        district: { type: string, nullable: true }
        rating: { type: number }
        reviewsCount: { type: integer }
        services:
          type: array
          items: { type: string }
        servicesAr:
          type: array
          items: { type: string }
        isVerified: { type: boolean }
        imageUrl: { type: string, nullable: true }
        priceLevel:
          type: string
          enum: [budget, standard, premium]
    WorkshopDetail:
      allOf:
        - $ref: "#/components/schemas/Workshop"
        - type: object
          properties:
            description: { type: string, nullable: true }
            descriptionAr: { type: string, nullable: true }
            openingHours: { type: string, nullable: true }
            address: { type: string, nullable: true }
            addressAr: { type: string, nullable: true }
            recentBookingsCount: { type: integer }
...
    InterpretDtcBody:
      type: object
      required: [code, vehicleId]
      properties:
        code: { type: string }
        vehicleId: { type: string }
    DtcInterpretation:
      type: object
      required: [descriptionAr, severity, action]
      properties:
        code: { type: string }
        descriptionAr: { type: string }
        severity:
          type: string
          enum: [low, medium, high, critical]
        causes:
          type: array
          items: { type: string }
        estimatedCostMin: { type: integer }
        estimatedCostMax: { type: integer }
        action:
          type: string
          enum: [drive_now, schedule_week, monitor]
        actionReasonAr: { type: string }

    # ===== Subscriptions =====
...